Fractional Knapsack
A greedy algorithm always makes the choice that looks best at the moment

That is, it makes a locally optimal choice in the hope that this choice will lead to a globally optimal solution
In the fractional knapsack problem, the setup is the same, but the thief can take fractions of items, rather than having to make a binary (0-1) choice for each item.

Think of an item in the 0-1 knapsack problem as being like a gold ingot, while an item in the fractional knapsack problem is more like gold dust.